+++ spec/test_optional_rule_component_tree.c
@@ -16,7 +16,7 @@ int main()
     assert(start->pR3 == NULL);
     assert(start->pR == NULL);
 
-    p_free_tree(start);
+    p_tree_delete(start);
     p_context_delete(context);
 
     input = "abcd";
@@ -31,7 +31,7 @@ int main()
     assert(start->pR == start->pR3);
     assert_eq(TOKEN_c, start->pR->pToken1->token);
 
-    p_free_tree(start);
+    p_tree_delete(start);
     p_context_delete(context);
 
     input = "bdc";
@@ -43,7 +43,7 @@ int main()
     assert(start->pR != NULL);
     assert_eq(TOKEN_d, start->pR->pToken1->token);
 
-    p_free_tree(start);
+    p_tree_delete(start);
     p_context_delete(context);
 
     return 0;